Refactor coin balance resolvers and structure
- Updated the coin balance resolvers to improve organization by moving them into a dedicated `resolvers` directory.
- Introduced a new `resolver.ts` file to define the `CoinBalanceResolver` type, enhancing type safety and clarity.
- Adjusted imports across the codebase to align with the new structure, ensuring consistency and maintainability.
- Implemented individual resolver functions for various chains, including Cardano, Cosmos, EVM, Polkadot, Ripple, Solana, Sui, Ton, Tron, and UTXO, to streamline balance retrieval logic.
