+++ core/chain/coin/balance/resolvers/evm.ts
@@ -1,9 +1,9 @@
 import { EvmChain } from '@core/chain/Chain'
+import { getEvmClient } from '@core/chain/chains/evm/client'
+import { getErc20Balance } from '@core/chain/chains/evm/erc20/getErc20Balance'
 import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
 
-import { getEvmClient } from '../../chains/evm/client'
-import { getErc20Balance } from '../../chains/evm/erc20/getErc20Balance'
-import { CoinBalanceResolver } from './CoinBalanceResolver'
+import { CoinBalanceResolver } from '../resolver'
 
 export const getEvmCoinBalance: CoinBalanceResolver<EvmChain> = async input => {
   return isFeeCoin(input)
