+++ core/chain/coin/balance/resolvers/ripple.ts
@@ -1,10 +1,10 @@
+import { getRippleAccountInfo } from '@core/chain/chains/ripple/account/info'
+import { getRippleNetworkInfo } from '@core/chain/chains/ripple/network/info'
 import { shouldBePresent } from '@lib/utils/assert/shouldBePresent'
 import { attempt } from '@lib/utils/attempt'
 import { isInError } from '@lib/utils/error/isInError'
 
-import { getRippleAccountInfo } from '../../chains/ripple/account/info'
-import { getRippleNetworkInfo } from '../../chains/ripple/network/info'
-import { CoinBalanceResolver } from './CoinBalanceResolver'
+import { CoinBalanceResolver } from '../resolver'
 
 export const getRippleCoinBalance: CoinBalanceResolver = async input => {
   const [accountResult, networkResult] = await Promise.all([
