+++ core/chain/coin/balance/resolvers/solana.ts
@@ -1,9 +1,9 @@
+import { getSolanaClient } from '@core/chain/chains/solana/client'
+import { getSplAccounts } from '@core/chain/chains/solana/spl/getSplAccounts'
 import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
 import { Address } from '@solana/web3.js'
 
-import { getSolanaClient } from '../../chains/solana/client'
-import { getSplAccounts } from '../../chains/solana/spl/getSplAccounts'
-import { CoinBalanceResolver } from './CoinBalanceResolver'
+import { CoinBalanceResolver } from '../resolver'
 
 export const getSolanaCoinBalance: CoinBalanceResolver = async input => {
   const client = getSolanaClient()
