+++ core/ui/chain/coin/queries/useBalancesQuery.ts
@@ -1,5 +1,5 @@
 import { getCoinBalance } from '@core/chain/coin/balance'
-import { CoinBalanceResolverInput } from '@core/chain/coin/balance/CoinBalanceResolver'
+import { CoinBalanceResolverInput } from '@core/chain/coin/balance/resolver'
 import { coinKeyToString } from '@core/chain/coin/Coin'
 import { useQueriesToEagerQuery } from '@lib/ui/query/hooks/useQueriesToEagerQuery'
 import { mergeRecords } from '@lib/utils/record/mergeRecords'
